requirements.txt
asgiref==3.8.1
Django==5.0.6
django-rest-framework==0.1.0
django-sslserver==0.22
djangorestframework==3.15.1
importlib==1.0.4
mysql-connector-python==8.4.0
mysqlclient==2.2.4
setuptools==70.0.0
sqlparse==0.5.0
tzdata==2024.1
